third_party/blink/web_tests/platform/generic/fast/forms/color/color-value-sanitization-expected.txt

CONSOLE WARNING: The specified value "" does not conform to the required format.  The format is "#rrggbb" where rr, gg, bb are two-digit hexadecimal numbers.
CONSOLE WARNING: The specified value "#ghijkl" does not conform to the required format.  The format is "#rrggbb" where rr, gg, bb are two-digit hexadecimal numbers.
CONSOLE WARNING: The specified value "#112233ff" does not conform to the required format.  The format is "#rrggbb" where rr, gg, bb are two-digit hexadecimal numbers.
CONSOLE WARNING: The specified value "ffccaa" does not conform to the required format.  The format is "#rrggbb" where rr, gg, bb are two-digit hexadecimal numbers.
CONSOLE WARNING: The specified value " #223344 " does not conform to the required format.  The format is "#rrggbb" where rr, gg, bb are two-digit hexadecimal numbers.
Value sanitization test for input[type=color]

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS input.value = "#000000"; input.value is "#000000"
PASS input.value = "#ABCDEF"; input.value is "#abcdef"
PASS input.value = "#abcdef"; input.value is "#abcdef"
PASS input.value = ""; input.value is "#000000"
PASS input.value = "#ghijkl"; input.value is "#000000"
PASS input.value = "#112233ff"; input.value is "#000000"
PASS input.value = "ffccaa"; input.value is "#000000"
PASS input.value = " #223344 "; input.value is "#000000"
PASS successfullyParsed is true

TEST COMPLETE
